back/requirements.txt
flask==2.3.2
python-dotenv==1.0.0
pymongo==4.3.3
boto3==1.26.142
flask-jwt-extended==4.5.2